Be simpler and closer to GMS2 behavior, just lex integers as i64
This does not allow for writing i64::MIN as a literal, but this is the
sanest thing to do for now without further language changes.

Any integer literal greater than i64::MAX (without any negation, as the
lexer does not consider a negative to be a single literal) will now lex
as a float.

The actual behavior of GMS2 for large integer constants is so bizarre
I can't fully explain it here.

Integral expressions (including any constant folding with negation,
addition, and subtaction) that fit into an i32 become *number* type (not
int32!). If those integral expressions do not fit into i32 they become
an int64 type instead.

Expressions that evaluate to near i64::MAX or i64::MIN behave extremely
strangely, becoming *clamped* at i64::MAX or i64::MIN even if they
are larger than can fit into an i64.

As an example, the literal `9223372036854775807` has type int64 and
value `9223372036854775807` in GMS2 as you would expect. However
`9223372036854775808` still has value `9223372036854775807`, as well
as `9223372036854775809` and `9223372036854775810` and a thousand or so
integers after that, until eventually becoming a number with the value
9223372036854777856.00. This makes a *kind* of sense in that the f64
epsilon at these values is very large, so for example there is no f64
with the value `9223372036854775808`, but it makes much more sense for
such a literal to either be an *error* or become the closest f64 value.

Additionally, literals near i64::MIN behave even more strangely in a
way that I can't even imagine how it is implemented. Here is a table of
typed expressions and the resulting values:

`-9223372036854775805` => `-9223372036854775805`
`-9223372036854775806` => `-9223372036854775806`
`-9223372036854775807` => `-9223372036854775808`
`-9223372036854775808` => `-9223372036854775808`
`-9223372036854775809` => `-9223372036854775808`
`-9223372036854775810` => `-9223372036854775808`

and so on, until getting close to a whole f64 value similarly to the
positive case.

Why `-i64::MAX` becomes `i64::MIN` I cannot fathom.

I have no intention of mimicking any of this behavior. It would be nice
to allow for i64::MIN to be able to be typed directly into source code,
but it feels very wrong in a language with no literal type suffixes and
no errors on out of range literals (Rust has both of these) to make the
following values of `n` have different types:

```
var n = -9223372036854775808;
```

vs

```
var n = -(9223372036854775808);
```

vs

```
var i = 9223372036854775808;
var n = -i;
```

So for now, do the simple, consistent thing.

+++ crates/compiler/src/ir_gen.rs
@@ -514,7 +514,7 @@ where
                         span,
                         ir::Instruction::SetIndexConst {
                             array,
-                            index: Constant::Integer(i as i128),
+                            index: Constant::Integer(i as i64),
                             value,
                         },
                     );
